My Money

+++ src/main/java/africa/semicolon/dto/requests/DepositRequest.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,7 +11,7 @@ import java.math.BigDecimal;
 public class DepositRequest {
     private String firstName;
     private String lastName;
-    private BigDecimal amount;
+    private int amount;
     private String accountNumber;
     private String email;
 }
